requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
distlib==0.3.9
Django==5.2.1
djangorestframework
python-dotenv
filelock==3.17.0
html5lib==1.1
idna==3.10
lxml==6.0.0
mysqlclient==2.2.7
oscrypto==1.3.0
pillow==11.3.0
platformdirs==4.3.6
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.9.0
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.29.3
webencodings==0.5.1
xhtml2pdf==0.2.17
pandas
matplotlib
seaborn
django-cors-headers==4.8.0